// apps/dashboard/src/server/db/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UsageQueueInboxStatus = {
  pending: 'pending',
  processing: 'processing',
  processed: 'processed',
  decode_failed: 'decode_failed',
  process_failed: 'process_failed',
  discarded: 'discarded'
} as const

export type UsageQueueInboxStatus = (typeof UsageQueueInboxStatus)[keyof typeof UsageQueueInboxStatus]